Added watch-website-changes
This script takes a URL as parameter, then retrieves the given page and
generates the MD5 sum of it. It then checks if the MD5 sum differs from the sum
generated on earlier calls of the script and echoes a message to stdout if the
page has changed since then.

I particularly use this to watch a change log from a project I'm interested in.